Made the following modifications: 1. Converted ScalarTransportEquation to template form so that it can be used for both staggered and non-staggered scalar fields. 2. Added proper functionality to the Wasatch problem specification to accommodate for staggered scalar transport. 3. Converted ScalarRHS to template form so that it conforms to other expressions that we have (i.e. DiffusiveFlux …). 4. Fixed wasatch input files.

+++ src/CCA/Components/Wasatch/Wasatch.cc
@@ -364,7 +364,7 @@ namespace Wasatch{
   {
     const double deltat = 1.0; // jcs should get this from an input file possibly?
 
-    for( int im=0; im<matls->size(); ++im ){
+    for( size_t im=0; im<matls->size(); ++im ){
 
       const int material = matls->get(im);
